change mysql to postgres (#7)
move assets from mysql to postgres

+++ core/eupower_core/eupower_core/dagster_resources/fs.py
@@ -63,7 +63,8 @@ class FsWriter(FsBase):
                 file.write(content)
 
     def delete_data(self):
-        shutil.rmtree(self.base_path)
+        if os.path.exists(self.base_path):
+            shutil.rmtree(self.base_path)
 
 
 class FsReader(FsBase):
